# src/api/swagger/api/openapi.yaml
openapi: 3.0.1
info:
  title: NS Checkin REST APIs Document
  # description: NS Checkin system REST API
  # termsOfService: http://swagger.io/terms/
  # contact:
  #   name: Hyperledger Team
  # license:
  #   name: Apache 2.0
  #   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://wns.novasquare.vn:8345
  - url: http://192.168.1.21:8345
paths:
...
  /api/country/getAll:
    get:
      tags:
        - countries
      summary: get all countries
      operationId: get all countries
      responses:
        "200":
          description: Success
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: UnAuthorized
          content: {}
        "404":
          description: Not found
          content: {}
      # security:
      #   - JWT: []
...
  /api/city/getAll:
    get:
      tags:
        - cities
      summary: get all cities
      operationId: get all cities
      responses:
        "200":
          description: Success
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: UnAuthorized
          content: {}
        "404":
          description: Not found
          content: {}
      security:
        - JWT: []
...
components:
...
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization